@@ -22,123 +22,117 @@
     "fetch:all": "npm run fetch:species:data && npm run fetch:species:profiles && npm run fetch:species:commonNames"
   },
   "dependencies": {
-    "@capacitor-community/background-geolocation": "^1.2.14",
-    "@capacitor/android": "^5.0.3",
-    "@capacitor/app": "^5.0.2",
-    "@capacitor/camera": "^5.0.2",
-    "@capacitor/core": "^5.0.3",
-    "@capacitor/device": "^5.0.2",
-    "@capacitor/filesystem": "^5.0.2",
-    "@capacitor/geolocation": "^5.0.2",
-    "@capacitor/haptics": "^5.0.2",
-    "@capacitor/ios": "^5.0.3",
-    "@capacitor/keyboard": "^5.0.2",
-    "@capacitor/network": "^5.0.2",
-    "@capacitor/push-notifications": "^5.0.2",
-    "@capacitor/splash-screen": "^5.0.2",
-    "@capacitor/status-bar": "^5.0.2",
-    "@flumens/bigu": "^0.4.0",
-    "@flumens/ionic": "^1.27.0",
-    "@indicia-js/core": "^6.1.0",
-    "@indicia-js/remote": "^6.2.0",
+    "@capacitor-community/background-geolocation": "1.2.14",
+    "@capacitor/android": "5.2.1",
+    "@capacitor/app": "5.0.6",
+    "@capacitor/camera": "5.0.6",
+    "@capacitor/core": "5.2.1",
+    "@capacitor/device": "5.0.6",
+    "@capacitor/filesystem": "5.1.0",
+    "@capacitor/geolocation": "5.0.6",
+    "@capacitor/haptics": "5.0.6",
+    "@capacitor/ios": "5.2.1",
+    "@capacitor/keyboard": "5.0.6",
+    "@capacitor/network": "5.0.6",
+    "@capacitor/push-notifications": "5.0.6",
+    "@capacitor/splash-screen": "5.0.6",
+    "@capacitor/status-bar": "5.0.6",
+    "@flumens/bigu": "0.4.0",
+    "@flumens/ionic": "2.0.0-alpha.4",
     "@ionic-native/core": "5.36.0",
-    "@ionic-native/launch-review": "^5.36.0",
+    "@ionic-native/launch-review": "5.36.0",
     "@ionic-native/sqlite": "5.36.0",
-    "@ionic/core": "^6.3.8",
-    "@ionic/react": "^6.3.8",
-    "@ionic/react-router": "^6.3.8",
+    "@ionic/core": "7.1.3",
+    "@ionic/react": "7.1.3",
+    "@ionic/react-router": "7.1.3",
     "@mapbox/geojson-area": "0.2.2",
-    "@sentry/browser": "^6.19.7",
-    "@sentry/integrations": "^6.19.7",
-    "@turf/distance": "^6.5.0",
-    "@types/wellknown": "^0.5.4",
-    "axios": "0.27.2",
-    "clsx": "^1.2.1",
-    "cordova-launch-review": "^4.0.1",
-    "cordova-plugin-device": "^2.1.0",
-    "cordova-sqlite-storage": "^6.1.0",
-    "d3": "^7.8.4",
-    "deep-diff": "^1.0.2",
-    "formik": "2.2.9",
-    "framer-motion": "4.1.17",
-    "i18next": "21.4.0",
-    "ionicons": "^6.0.4",
-    "jwt-decode": "^3.1.2",
-    "leaflet": "^1.9.3",
-    "leaflet-draw": "1.0.4",
+    "@mapbox/mapbox-gl-draw": "1.4.2",
+    "@mapbox/sphericalmercator": "^1.2.0",
+    "@sentry/browser": "7.58.1",
+    "@sentry/integrations": "7.58.1",
+    "@turf/distance": "6.5.0",
+    "@turf/kinks": "^6.5.0",
+    "@types/wellknown": "0.5.5",
+    "axios": "1.4.0",
+    "clsx": "2.0.0",
+    "cordova-launch-review": "4.0.1",
+    "cordova-plugin-device": "2.1.0",
+    "cordova-sqlite-storage": "6.1.0",
+    "d3": "7.8.5",
+    "deep-diff": "1.0.2",
+    "formik": "2.4.2",
+    "i18next": "23.2.11",
+    "ionicons": "7.1.2",
+    "jwt-decode": "3.1.2",
     "localforage": "1.10.0",
     "localforage-cordovasqlitedriver": "1.8.0",
     "lodash": "4.17.21",
-    "mapbox-gl": "2.4.1",
-    "mobx": "^6.9.0",
-    "mobx-react": "^7.6.0",
-    "mobx-utils": "^6.0.6",
-    "proj4": "^2.9.0",
-    "proj4leaflet": "^1.0.2",
-    "prop-types": "^15.8.1",
-    "prop-types-exact": "^1.2.0",
-    "react": "17.0.2",
-    "react-autosuggest": "^10.1.0",
+    "mobx": "6.9.1",
+    "mobx-react": "7.6.0",
+    "mobx-utils": "6.0.7",
+    "prop-types": "15.8.1",
+    "prop-types-exact": "1.2.0",
+    "react": "18.2.0",
+    "react-autosuggest": "10.1.0",
     "react-countdown-now": "2.1.2",
-    "react-dom": "17.0.2",
-    "react-i18next": "11.13.0",
-    "react-leaflet": "3.2.2",
-    "react-leaflet-markercluster": "^3.0.0-rc1",
-    "react-router": "5.2.1",
-    "react-router-dom": "5.3.0",
-    "react-window": "^1.8.9",
-    "suncalc": "^1.9.0",
-    "swiper": "^8.4.5",
-    "wellknown": "^0.5.0",
-    "yup": "0.32.11"
+    "react-dom": "18.2.0",
+    "react-i18next": "13.0.2",
+    "react-map-gl": "^7.1.2",
+    "react-router": "5.3.4",
+    "react-router-dom": "5.3.4",
+    "react-window": "1.8.9",
+    "suncalc": "1.9.0",
+    "swiper": "10.0.4",
+    "wellknown": "0.5.0",
+    "yup": "1.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.21.4",
+    "@babel/core": "7.21.4",
     "@babel/plugin-syntax-dynamic-import": "7.8.3",
-    "@babel/plugin-transform-exponentiation-operator": "^7.18.6",
-    "@babel/plugin-transform-runtime": "^7.21.4",
-    "@babel/preset-env": "^7.21.4",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.21.4",
-    "@capacitor/cli": "^5.0.3",
-    "@flumens/eslint-config": "^2.0.1",
-    "@flumens/fetch-onedrive-excel": "^0.3.0",
-    "@flumens/prettier-config": "^0.2.3",
-    "@flumens/webpack-config": "^5.0.0",
-    "@types/jest": "^27.5.1",
-    "@types/react": "^17.0.8",
-    "@types/react-dom": "^17.0.5",
-    "@types/react-leaflet": "^2.8.2",
-    "@types/react-leaflet-markercluster": "^3.0.1",
-    "@types/react-router": "^5.1.20",
-    "@types/react-router-dom": "^5.3.3",
-    "@types/react-window": "^1.8.5",
-    "@types/suncalc": "^1.9.0",
-    "babel-jest": "^28.1.0",
-    "babel-plugin-module-resolver": "^4.1.0",
-    "btoa": "^1.2.1",
-    "cordova-res": "^0.15.4",
+    "@babel/plugin-transform-exponentiation-operator": "7.18.6",
+    "@babel/plugin-transform-runtime": "7.21.4",
+    "@babel/preset-env": "7.21.4",
+    "@babel/preset-react": "7.18.6",
+    "@babel/preset-typescript": "7.21.4",
+    "@capacitor/cli": "5.0.3",
+    "@flumens/eslint-config": "2.0.1",
+    "@flumens/fetch-onedrive-excel": "0.3.0",
+    "@flumens/prettier-config": "0.2.3",
+    "@flumens/webpack-config": "5.0.0",
+    "@types/jest": "29.5.3",
+    "@types/lodash": "^4.14.195",
+    "@types/mapbox__sphericalmercator": "^1.2.0",
+    "@types/react": "18.2.15",
+    "@types/react-dom": "18.2.7",
+    "@types/react-router": "5.1.20",
+    "@types/react-router-dom": "5.3.3",
+    "@types/react-window": "1.8.5",
+    "@types/suncalc": "1.9.0",
+    "babel-jest": "29.6.1",
+    "babel-plugin-module-resolver": "5.0.0",
+    "btoa": "1.2.1",
+    "cordova-res": "0.15.4",
     "csvtojson": "2.0.10",
-    "dotenv": "^16.0.3",
-    "eslint": "^8.37.0",
-    "gettext-parser": "^2.0.0",
+    "dotenv": "16.0.3",
+    "eslint": "8.37.0",
+    "gettext-parser": "2.0.0",
     "gettext-to-messageformat": "0.3.1",
-    "husky": "^8.0.3",
-    "jest": "^28.1.0",
-    "jest-environment-jsdom": "^28.1.0",
-    "jest-transform-stub": "^2.0.0",
+    "husky": "8.0.3",
+    "jest": "29.6.1",
+    "jest-environment-jsdom": "29.6.1",
+    "jest-transform-stub": "2.0.0",
     "mock-geolocation": "1.0.11",
     "request": "2.88.2",
-    "tslib": "^2.5.0",
-    "typescript": "^4.9.3",
-    "vercel": "^24.2.4",
-    "webpack": "^5.77.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.13.2"
+    "tslib": "2.5.0",
+    "typescript": "5.1.6",
+    "vercel": "31.0.3",
+    "webpack": "5.77.0",
+    "webpack-cli": "5.1.4",
+    "webpack-dev-server": "4.13.2"
   },
   "engines": {
-    "node": ">=12",
-    "npm": ">=6"
+    "node": ">=16",
+    "npm": ">=9"
   },
   "author": {
     "name": "Flumens",
