@@ -1,5 +1,5 @@
 package uk.ac.ceh.ebms;
+
 import com.getcapacitor.BridgeActivity;
 
-public class MainActivity extends BridgeActivity {
-}
+public class MainActivity extends BridgeActivity {}
