@@ -1,188 +1,220 @@
-import { FC, useState } from 'react';
+import { useState, useEffect, useCallback } from 'react';
 import { observer } from 'mobx-react';
-import L, { LatLngExpression } from 'leaflet';
-import 'mapbox-gl';
-import 'mapbox-gl/dist/mapbox-gl.css';
-import { useTranslation } from 'react-i18next';
-import { MapContainer } from 'react-leaflet';
-import { date as dateHelp } from '@flumens';
-import '@flumens/ionic/dist/components/ModelLocationMap/Map/map/leaflet-mapbox-gl';
-import { useIonViewDidEnter } from '@ionic/react';
-import CONFIG from 'common/config';
-import hasWebGL from 'common/helpers/webGLSupport';
-import savedSamples from 'models/collections/samples';
-import Sample from 'models/sample';
+import { Trans as T } from 'react-i18next';
+import { MapRef, LngLatBounds } from 'react-map-gl';
+import { Link } from 'react-router-dom';
+import {
+  useToast,
+  device,
+  MapContainer,
+  useCenterMapToCurrentLocation,
+} from '@flumens';
+import { IonSpinner } from '@ionic/react';
+import config from 'common/config';
+import { centroids as countries } from 'common/config/countries';
+import appModel from 'common/models/app';
+import userModel from 'models/user';
+import GPS from 'helpers/GPS';
+import { Square, Record } from './esResponse.d';
+import { fetchRecords, fetchSquares } from './recordsService';
 import './styles.scss';
 
-const DEFAULT_ZOOM = 5;
-const DEFAULT_CENTER: LatLngExpression = [51.505, -0.09];
+/**
+ * Returns square size in meters.
+ */
+const getSquareSize = (zoomLevel: number) => {
+  if (zoomLevel < 8) return 10000;
+  if (zoomLevel < 10) return 2000;
 
-const URL =
-  'https://api.mapbox.com/styles/v1/cehapps/cipqvo0c0000jcknge1z28ejp/tiles/256/{z}/{x}/{y}?access_token={accessToken}';
-
-const MapBoxAttribution =
-  '<a href="http://mapbox.com/about/maps" class="mapbox-wordmark" target="_blank">Mapbox</a><input type="checkbox" id="toggle-info"> <label for="toggle-info"><img src="data:image/svg+xml;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iaXNvLTg4NTktMSI/Pg0KPCEtLSBHZW5lcmF0b3I6IEFkb2JlIElsbHVzdHJhdG9yIDE4LjEuMSwgU1ZHIEV4cG9ydCBQbHVnLUluIC4gU1ZHIFZlcnNpb246IDYuMDAgQnVpbGQgMCkgIC0tPg0KPHN2ZyB2ZXJzaW9uPSIxLjEiIGlkPSJDYXBhXzEiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgeG1sbnM6eGxpbms9Imh0dHA6Ly93d3cudzMub3JnLzE5OTkveGxpbmsiIHg9IjBweCIgeT0iMHB4Ig0KCSB2aWV3Qm94PSIwIDAgNDIyLjY4NiA0MjIuNjg2IiBzdHlsZT0iZW5hYmxlLWJhY2tncm91bmQ6bmV3IDAgMCA0MjIuNjg2IDQyMi42ODY7IiB4bWw6c3BhY2U9InByZXNlcnZlIj4NCjxnPg0KCTxnPg0KCQk8cGF0aCBzdHlsZT0iZmlsbDojMDEwMDAyOyIgZD0iTTIxMS4zNDMsNDIyLjY4NkM5NC44MDQsNDIyLjY4NiwwLDMyNy44ODIsMCwyMTEuMzQzQzAsOTQuODEyLDk0LjgxMiwwLDIxMS4zNDMsMA0KCQkJczIxMS4zNDMsOTQuODEyLDIxMS4zNDMsMjExLjM0M0M0MjIuNjg2LDMyNy44ODIsMzI3Ljg4Miw0MjIuNjg2LDIxMS4zNDMsNDIyLjY4NnogTTIxMS4zNDMsMTYuMjU3DQoJCQljLTEwNy41NzQsMC0xOTUuMDg2LDg3LjUyLTE5NS4wODYsMTk1LjA4NnM4Ny41MiwxOTUuMDg2LDE5NS4wODYsMTk1LjA4NnMxOTUuMDg2LTg3LjUyLDE5NS4wODYtMTk1LjA4Ng0KCQkJUzMxOC45MDgsMTYuMjU3LDIxMS4zNDMsMTYuMjU3eiIvPg0KCTwvZz4NCgk8Zz4NCgkJPGc+DQoJCQk8cGF0aCBzdHlsZT0iZmlsbDojMDEwMDAyOyIgZD0iTTIzMS45LDEwNC42NDdjMC4zNjYsMTEuMzIzLTcuOTM0LDIwLjM3LTIxLjEzNCwyMC4zN2MtMTEuNjg5LDAtMTkuOTk2LTkuMDU1LTE5Ljk5Ni0yMC4zNw0KCQkJCWMwLTExLjY4OSw4LjY4MS0yMC43NDQsMjAuNzQ0LTIwLjc0NEMyMjMuOTc1LDgzLjkwMywyMzEuOSw5Mi45NTgsMjMxLjksMTA0LjY0N3ogTTE5NC45MzEsMzM4LjUzMVYxNTUuOTU1aDMzLjE4OXYxODIuNTc2DQoJCQkJQzIyOC4xMiwzMzguNTMxLDE5NC45MzEsMzM4LjUzMSwxOTQuOTMxLDMzOC41MzF6Ii8+DQoJCTwvZz4NCgk8L2c+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8L3N2Zz4NCg==" /></label> <div>Leaflet © <a href="https://www.mapbox.com/about/maps/">Mapbox</a> © <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a> <strong><a href="https://www.mapbox.com/map-feedback/" target="_blank">Improve this map</a></strong></div>';
-
-function getTransectMarker(sample: Sample) {
-  let latitude;
-  let longitude;
-  try {
-    const { centroid_sref: centroid } = sample.attrs.location || {};
-    [latitude, longitude] = centroid.split(' ').map(parseFloat);
-  } catch (e) {
-    return null;
-  }
-
-  const square = L.divIcon({
-    className: 'square-leaflet-icon',
-  });
-
-  return L.marker([latitude, longitude], { icon: square });
-}
-
-function getAreaCountMarker(sample: Sample) {
-  const { latitude, longitude } = sample.attrs.location || {};
-
-  if (!latitude) {
-    return null;
-  }
-
-  return L.circleMarker([latitude, longitude], {
-    color: 'white',
-    fillColor: '#745a8f',
-    fillOpacity: 1,
-    weight: 4,
-  });
-}
-
-function getTransectPosition(sample: Sample) {
-  try {
-    // wrapping in try/catch because can't control the centroid_sref format
-    const { centroid_sref: centroid } = sample.attrs.location || {};
-    const [latitude, longitude] = centroid.split(' ').map(parseFloat);
-    return [latitude, longitude];
-  } catch (e) {
-    return [];
-  }
-}
-
-function getAreaCountPosition(sample: Sample) {
-  const { latitude, longitude } = sample.attrs.location || {};
-  if (!latitude) {
-    return [];
-  }
-
-  return [latitude, longitude];
-}
-
-type Props = {
-  showingMap?: boolean;
+  return 1000;
 };
 
-const Component: FC<Props> = () => {
-  const [map, setMap]: any = useState(null);
+const getTotalSquares = (squares: Square[]) => {
+  const addSquares = (acc: number, square: Square): number =>
+    acc + square.doc_count;
 
-  const { t } = useTranslation();
+  // protection division from 0, defaulting to 1
+  return squares?.reduce(addSquares, 0) || 1;
+};
 
-  const addRecordsPopup = (sample: Sample, marker: any) => {
-    const date = dateHelp.print(sample.attrs.date, true);
+const Map = () => {
+  const [mapRef, setMapRef] = useState<{ current?: MapRef }>({});
+  const measuredRef = useCallback(
+    (node: any) => node && setMapRef({ current: node }),
+    []
+  );
 
-    let speciesInfo = '';
-    speciesInfo = !sample.metadata.syncedOn
-      ? `<br/>(<i>${t('Pending')}</i>)`
-      : '';
+  const [isFetchingRecords, setFetchingRecords] = useState<any>(null);
+  const toast = useToast();
 
-    if (sample.metadata.survey !== 'transect') {
-      speciesInfo += `<br/>${t('Species')}: ${sample.samples.length}`;
-    }
+  const [totalSquares, setTotalSquares] = useState<number>(1);
+  const [squares, setSquares] = useState<Square[]>([]);
+  const [records, setRecords] = useState<Record[]>([]);
 
-    marker.bindPopup(`<b>${date}</b>${speciesInfo}`);
-  };
+  const userIsLoggedIn = userModel.isLoggedIn();
 
-  const addRecords = (mapRef: L.Map) => {
-    const addSurveyMarkerToMap = (sample: Sample) => {
-      const marker =
-        sample.metadata.survey === 'transect'
-          ? getTransectMarker(sample)
-          : getAreaCountMarker(sample);
+  const updateRecords = async () => {
+    if (
+      !mapRef.current ||
+      !userIsLoggedIn ||
+      !userModel.attrs.verified ||
+      !device.isOnline
+    )
+      return;
 
-      if (!marker) {
-        return;
-      }
+    const bounds: LngLatBounds = mapRef.current.getBounds(); // TODO: .pad(0.5); // padding +50%
 
-      addRecordsPopup(sample, marker);
-      marker.addTo(mapRef);
-    };
+    const zoomLevel = mapRef.current.getZoom();
+    const northWest = bounds.getNorthWest();
+    const southEast = bounds.getSouthEast();
 
-    savedSamples.forEach(addSurveyMarkerToMap);
-  };
+    if (northWest.lat === southEast.lat) return; // first time the bounds can be flat
 
-  const zoomToRecords = (mapRef: L.Map) => {
-    const getPositionsFromSurveys = (agg: any, sample: Sample) => {
-      const position =
-        sample.metadata.survey === 'transect'
-          ? getTransectPosition(sample)
-          : getAreaCountPosition(sample);
-
-      if (!position.length) {
-        return agg;
-      }
-
-      return [...agg, position];
-    };
-
-    const positions = savedSamples.reduce(getPositionsFromSurveys, []);
-
-    if (!positions.length) {
+    const shouldFetchRecords = zoomLevel >= 13;
+    if (shouldFetchRecords) {
+      setFetchingRecords(true);
+      const fetchedRecords = await fetchRecords(northWest, southEast).catch(
+        toast.error
+      );
+      // Previous request was cancelled
+      if (!fetchedRecords) return;
+      setRecords(fetchedRecords);
+      setSquares([]);
+      setFetchingRecords(false);
       return;
     }
 
-    mapRef.fitBounds(positions);
+    const squareSize = getSquareSize(zoomLevel);
 
-    // TODO: BUG! Single position or multiple positions but same coords(like a transect) have same bounds and zoom becames infinity
-    const zoom = mapRef.getZoom();
-    if (zoom === Infinity) {
-      mapRef.setZoom(DEFAULT_ZOOM);
-    }
+    setFetchingRecords(true);
+    const fetchedSquares = await fetchSquares(
+      northWest,
+      southEast,
+      squareSize
+    ).catch(toast.error);
+
+    // Previous request was cancelled
+    if (!fetchedSquares) return;
+    setRecords([]);
+
+    setTotalSquares(getTotalSquares(fetchedSquares));
+    setSquares(fetchedSquares);
+    setFetchingRecords(false);
   };
 
-  const initMap = (mapRef: L.Map) => {
-    setMap(mapRef);
+  const updateMapCentre = () => updateRecords();
 
-    const suppportsWebGL = hasWebGL();
+  const updateRecordsFirstTime = () => {
+    updateRecords();
+  };
+  useEffect(updateRecordsFirstTime, [mapRef]);
 
-    const layer = suppportsWebGL
-      ? (L as any).mapboxGL({
-          accessToken: CONFIG.map.mapboxApiKey,
-          style: 'mapbox://styles/mapbox/satellite-streets-v11',
-          attribution: MapBoxAttribution,
-        })
-      : L.tileLayer(URL, {
-          attribution:
-            '<a href="http://mapbox.com/about/maps" class="mapbox-wordmark" target="_blank">Mapbox</a><input type="checkbox" id="toggle-info"> <label for="toggle-info"><img src="data:image/svg+xml;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iaXNvLTg4NTktMSI/Pg0KPCEtLSBHZW5lcmF0b3I6IEFkb2JlIElsbHVzdHJhdG9yIDE4LjEuMSwgU1ZHIEV4cG9ydCBQbHVnLUluIC4gU1ZHIFZlcnNpb246IDYuMDAgQnVpbGQgMCkgIC0tPg0KPHN2ZyB2ZXJzaW9uPSIxLjEiIGlkPSJDYXBhXzEiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgeG1sbnM6eGxpbms9Imh0dHA6Ly93d3cudzMub3JnLzE5OTkveGxpbmsiIHg9IjBweCIgeT0iMHB4Ig0KCSB2aWV3Qm94PSIwIDAgNDIyLjY4NiA0MjIuNjg2IiBzdHlsZT0iZW5hYmxlLWJhY2tncm91bmQ6bmV3IDAgMCA0MjIuNjg2IDQyMi42ODY7IiB4bWw6c3BhY2U9InByZXNlcnZlIj4NCjxnPg0KCTxnPg0KCQk8cGF0aCBzdHlsZT0iZmlsbDojMDEwMDAyOyIgZD0iTTIxMS4zNDMsNDIyLjY4NkM5NC44MDQsNDIyLjY4NiwwLDMyNy44ODIsMCwyMTEuMzQzQzAsOTQuODEyLDk0LjgxMiwwLDIxMS4zNDMsMA0KCQkJczIxMS4zNDMsOTQuODEyLDIxMS4zNDMsMjExLjM0M0M0MjIuNjg2LDMyNy44ODIsMzI3Ljg4Miw0MjIuNjg2LDIxMS4zNDMsNDIyLjY4NnogTTIxMS4zNDMsMTYuMjU3DQoJCQljLTEwNy41NzQsMC0xOTUuMDg2LDg3LjUyLTE5NS4wODYsMTk1LjA4NnM4Ny41MiwxOTUuMDg2LDE5NS4wODYsMTk1LjA4NnMxOTUuMDg2LTg3LjUyLDE5NS4wODYtMTk1LjA4Ng0KCQkJUzMxOC45MDgsMTYuMjU3LDIxMS4zNDMsMTYuMjU3eiIvPg0KCTwvZz4NCgk8Zz4NCgkJPGc+DQoJCQk8cGF0aCBzdHlsZT0iZmlsbDojMDEwMDAyOyIgZD0iTTIzMS45LDEwNC42NDdjMC4zNjYsMTEuMzIzLTcuOTM0LDIwLjM3LTIxLjEzNCwyMC4zN2MtMTEuNjg5LDAtMTkuOTk2LTkuMDU1LTE5Ljk5Ni0yMC4zNw0KCQkJCWMwLTExLjY4OSw4LjY4MS0yMC43NDQsMjAuNzQ0LTIwLjc0NEMyMjMuOTc1LDgzLjkwMywyMzEuOSw5Mi45NTgsMjMxLjksMTA0LjY0N3ogTTE5NC45MzEsMzM4LjUzMVYxNTUuOTU1aDMzLjE4OXYxODIuNTc2DQoJCQkJQzIyOC4xMiwzMzguNTMxLDE5NC45MzEsMzM4LjUzMSwxOTQuOTMxLDMzOC41MzF6Ii8+DQoJCTwvZz4NCgk8L2c+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8L3N2Zz4NCg==" /></label> <div>Leaflet © <a href="https://www.mapbox.com/about/maps/">Mapbox</a> © <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a> <strong><a href="https://www.mapbox.com/map-feedback/" target="_blank">Improve this map</a></strong></div>',
-          accessToken: CONFIG.map.mapboxApiKey,
-          tileSize: 256, // specify as, OS layer overwites this with 200 otherwise,
-          minZoom: DEFAULT_ZOOM,
-        });
+  const getRecordMarker = (record: Record) => {
+    const [latitude, longitude] = record.location.point
+      .split(',')
+      .map(parseFloat);
 
-    layer.addTo(mapRef);
+    return (
+      <MapContainer.Marker.Circle
+        key={record.id}
+        id={record.id}
+        longitude={longitude}
+        latitude={latitude}
+        paint={{
+          'circle-stroke-color': 'white',
+          'circle-color': '#745a8f',
+          'circle-opacity': 1,
+        }}
+      />
+    );
+  };
+  const recordMarkers = records.map(getRecordMarker);
 
-    addRecords(mapRef);
+  const getSquareMarker = (square: Square) => {
+    const opacity = Number((square.doc_count / totalSquares).toFixed(2));
 
-    zoomToRecords(mapRef);
+    // max 80%, min 20%
+    const normalizedOpacity = Math.min(Math.max(opacity, 0.2), 0.8);
+
+    const [longitude, latitude] = square.key.split(' ').map(parseFloat);
+
+    const radius = square.size! / 2;
+    const padding = 1.1; // extra padding between squares
+    const metersToPixels =
+      radius / padding / 0.075 / Math.cos((latitude * Math.PI) / 180);
+
+    return (
+      <MapContainer.Marker.Circle
+        key={square.key}
+        id={square.key}
+        longitude={longitude}
+        latitude={latitude}
+        paint={{
+          'circle-stroke-color': 'white',
+          'circle-color': '#745a8f',
+          'circle-opacity': normalizedOpacity,
+          'circle-radius': [
+            'interpolate',
+            ['exponential', 2],
+            ['zoom'],
+            0,
+            0,
+            20,
+            metersToPixels,
+          ],
+        }}
+      />
+    );
   };
 
-  const updateLayout = () => map && map.invalidateSize();
-  useIonViewDidEnter(updateLayout, [map]);
+  const squareMarkers = squares.map(getSquareMarker);
+
+  const { isLocating, centerMapToCurrentLocation } =
+    useCenterMapToCurrentLocation(GPS);
+
+  let initialViewState;
+  const country = countries[appModel.attrs.country];
+  if (country?.zoom) {
+    initialViewState = { ...country };
+  }
 
   return (
-    <MapContainer
-      id="surveys-map"
-      whenCreated={initMap}
-      center={DEFAULT_CENTER}
-      zoom={DEFAULT_ZOOM}
-    />
+    <>
+      <MapContainer
+        id="user-records"
+        ref={measuredRef}
+        accessToken={config.map.mapboxApiKey}
+        mapStyle="mapbox://styles/mapbox/satellite-streets-v10"
+        maxPitch={0}
+        maxZoom={20}
+        initialViewState={initialViewState}
+        onMoveEnd={updateMapCentre}
+      >
+        {!userIsLoggedIn && (
+          <div className="login-message">
+            <T>
+              You need to login to your{' '}
+              <Link to="/user/login">iRecord account</Link> to be able to view
+              the records.
+            </T>
+          </div>
+        )}
+
+        <MapContainer.Control.Geolocate
+          isLocating={isLocating}
+          onClick={centerMapToCurrentLocation}
+        />
+
+        {squareMarkers}
+
+        {recordMarkers}
+
+        <MapContainer.Control>
+          {isFetchingRecords ? <IonSpinner /> : <div />}
+        </MapContainer.Control>
+      </MapContainer>
+    </>
   );
 };
 
-export default observer(Component);
+export default observer(Map);
