@@ -1,240 +1,208 @@
-import { FC, useEffect, useState, useRef } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { observer } from 'mobx-react';
-import { starOutline } from 'ionicons/icons';
-import L, { LatLngExpression } from 'leaflet';
-import 'leaflet-draw';
-import 'mapbox-gl';
-import 'mapbox-gl/dist/mapbox-gl.css';
-import { MapContainer, Marker } from 'react-leaflet';
-import { useRouteMatch } from 'react-router';
-import { Haptics, ImpactStyle } from '@capacitor/haptics';
-import { Main } from '@flumens';
-import '@flumens/ionic/dist/components/ModelLocationMap/Map/map/leaflet-mapbox-gl';
+import { MapRef } from 'react-map-gl';
 import {
-  IonToolbar,
-  IonTitle,
-  isPlatform,
-  IonIcon,
-  IonModal,
-  IonContent,
-  useIonViewDidEnter,
-  useIonViewWillLeave,
-  IonHeader,
-} from '@ionic/react';
-import MapControl from 'common/Components/LeafletControl';
-import CONFIG from 'common/config';
-import hasWebGL from 'common/helpers/webGLSupport';
+  Main,
+  MapContainer,
+  MapDraw,
+  useAlert,
+  useCenterMapToCurrentLocation,
+  CircleMarker,
+} from '@flumens';
+import config from 'common/config';
 import Sample from 'models/sample';
-import PastLocationsList from '../common/Components/PastLocationsList';
-import MapInfo from './MapInfo';
+import GPS from 'helpers/GPS';
+import Favourites from './Favourites';
+import Records from './Records';
 
-const SNAP_POSITIONS = [0, 0.3, 0.5, 1];
-const DEFAULT_SNAP_POSITION = 0.3;
+const useDeletePropt = () => {
+  const alert = useAlert();
 
-const MapBoxAttribution =
-  '<a href="http://mapbox.com/about/maps" class="mapbox-wordmark" target="_blank">Mapbox</a><input type="checkbox" id="toggle-info"> <label for="toggle-info"><img src="data:image/svg+xml;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iaXNvLTg4NTktMSI/Pg0KPCEtLSBHZW5lcmF0b3I6IEFkb2JlIElsbHVzdHJhdG9yIDE4LjEuMSwgU1ZHIEV4cG9ydCBQbHVnLUluIC4gU1ZHIFZlcnNpb246IDYuMDAgQnVpbGQgMCkgIC0tPg0KPHN2ZyB2ZXJzaW9uPSIxLjEiIGlkPSJDYXBhXzEiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgeG1sbnM6eGxpbms9Imh0dHA6Ly93d3cudzMub3JnLzE5OTkveGxpbmsiIHg9IjBweCIgeT0iMHB4Ig0KCSB2aWV3Qm94PSIwIDAgNDIyLjY4NiA0MjIuNjg2IiBzdHlsZT0iZW5hYmxlLWJhY2tncm91bmQ6bmV3IDAgMCA0MjIuNjg2IDQyMi42ODY7IiB4bWw6c3BhY2U9InByZXNlcnZlIj4NCjxnPg0KCTxnPg0KCQk8cGF0aCBzdHlsZT0iZmlsbDojMDEwMDAyOyIgZD0iTTIxMS4zNDMsNDIyLjY4NkM5NC44MDQsNDIyLjY4NiwwLDMyNy44ODIsMCwyMTEuMzQzQzAsOTQuODEyLDk0LjgxMiwwLDIxMS4zNDMsMA0KCQkJczIxMS4zNDMsOTQuODEyLDIxMS4zNDMsMjExLjM0M0M0MjIuNjg2LDMyNy44ODIsMzI3Ljg4Miw0MjIuNjg2LDIxMS4zNDMsNDIyLjY4NnogTTIxMS4zNDMsMTYuMjU3DQoJCQljLTEwNy41NzQsMC0xOTUuMDg2LDg3LjUyLTE5NS4wODYsMTk1LjA4NnM4Ny41MiwxOTUuMDg2LDE5NS4wODYsMTk1LjA4NnMxOTUuMDg2LTg3LjUyLDE5NS4wODYtMTk1LjA4Ng0KCQkJUzMxOC45MDgsMTYuMjU3LDIxMS4zNDMsMTYuMjU3eiIvPg0KCTwvZz4NCgk8Zz4NCgkJPGc+DQoJCQk8cGF0aCBzdHlsZT0iZmlsbDojMDEwMDAyOyIgZD0iTTIzMS45LDEwNC42NDdjMC4zNjYsMTEuMzIzLTcuOTM0LDIwLjM3LTIxLjEzNCwyMC4zN2MtMTEuNjg5LDAtMTkuOTk2LTkuMDU1LTE5Ljk5Ni0yMC4zNw0KCQkJCWMwLTExLjY4OSw4LjY4MS0yMC43NDQsMjAuNzQ0LTIwLjc0NEMyMjMuOTc1LDgzLjkwMywyMzEuOSw5Mi45NTgsMjMxLjksMTA0LjY0N3ogTTE5NC45MzEsMzM4LjUzMVYxNTUuOTU1aDMzLjE4OXYxODIuNTc2DQoJCQkJQzIyOC4xMiwzMzguNTMxLDE5NC45MzEsMzM4LjUzMSwxOTQuOTMxLDMzOC41MzF6Ii8+DQoJCTwvZz4NCgk8L2c+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8Zz4NCjwvZz4NCjxnPg0KPC9nPg0KPGc+DQo8L2c+DQo8L3N2Zz4NCg==" /></label> <div>Leaflet © <a href="https://www.mapbox.com/about/maps/">Mapbox</a> © <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a> <strong><a href="https://www.mapbox.com/map-feedback/" target="_blank">Improve this map</a></strong></div>';
-
-const URL =
-  'https://api.mapbox.com/styles/v1/cehapps/cipqvo0c0000jcknge1z28ejp/tiles/256/{z}/{x}/{y}?access_token={accessToken}';
-
-L.Icon.Default.imagePath = '/images';
-
-const finishFlag = L.divIcon({
-  className: 'finish-flag',
-  html: `<span />`,
-  iconAnchor: [11, 27],
-});
-
-const startFlag = L.divIcon({
-  className: 'start-flag',
-  html: `<span />`,
-  iconAnchor: [11, 27],
-});
-
-const lastFlag = L.divIcon({
-  className: 'last-flag',
-  html: `<span />`,
-  iconAnchor: [11, 27],
-});
-
-const DEFAULT_ZOOM = 5;
+  return () =>
+    new Promise((resolve: any) => {
+      alert({
+        header: 'Delete',
+        message: 'Are you sure you want to delete your current track?',
+        buttons: [
+          {
+            text: 'Cancel',
+            role: 'cancel',
+            handler: () => resolve(false),
+          },
+          {
+            text: 'Delete',
+            role: 'destructive',
+            handler: () => resolve(true),
+          },
+        ],
+      });
+    });
+};
 
 type Props = {
   sample: Sample;
-  shape: any;
   setLocation: any;
-  areaPretty: any;
   isGPSTracking: boolean;
   isDisabled?: boolean;
 };
 
-const AreaAttr: FC<Props> = ({
+const AreaAttr = ({
   sample,
-  shape,
   setLocation,
-  areaPretty,
   isGPSTracking,
   isDisabled,
-}) => {
+}: Props) => {
+  const mapRef = useRef<MapRef>(null);
+  const { location } = sample.attrs;
+
+  const [mapCenter, saveMapCenter] = useState<any>([1, 1]);
+  const updateMapCentre = ({ viewState }: any) =>
+    saveMapCenter([viewState.latitude, viewState.longitude]);
+
   const [showPastLocations, setShowPastLocations] = useState(false);
-  const [currentMapCenter, setMapCurrentCenter] = useState([51, -1]);
 
-  const buttonRef = useRef<any>(null);
-  const fixFavBtnClickMapPropagation = () =>
-    buttonRef.current && L.DomEvent.disableClickPropagation(buttonRef.current);
-  useIonViewDidEnter(fixFavBtnClickMapPropagation);
-
-  const fixOpenModalWhenLeavingPage = () => setShowPastLocations(false);
-  useIonViewWillLeave(fixOpenModalWhenLeavingPage);
-
-  const onSelectPastLoaction = (location: any) => {
-    if (sample.isGPSRunning()) sample.stopGPS();
-    setShowPastLocations(false);
-
-    isPlatform('hybrid') && Haptics.impact({ style: ImpactStyle.Light });
-
-    // eslint-disable-next-line no-param-reassign
-    sample.attrs.location = location;
-    sample.save();
-  };
-
-  const disableTapForIOS = !isPlatform('ios'); // TODO: https://github.com/Leaflet/Leaflet/issues/7255
-
-  const [map, setMap]: any = useState(null);
-
-  const setTileLayer = (mapRef: L.Map) => {
-    const suppportsWebGL = hasWebGL();
-
-    if (suppportsWebGL) {
-      (L as any)
-        .mapboxGL({
-          accessToken: CONFIG.map.mapboxApiKey,
-          style: 'mapbox://styles/mapbox/satellite-streets-v11',
-          // eslint-disable-next-line
-          // @ts-ignore
-          attribution: MapBoxAttribution,
-        })
-        .addTo(mapRef);
-    } else {
-      L.tileLayer(URL, {
-        attribution: MapBoxAttribution,
-        accessToken: CONFIG.map.mapboxApiKey,
-      }).addTo(mapRef);
-    }
-
-    setMap(mapRef);
-  };
-
-  const match = useRouteMatch();
+  const shouldDeleteShape = useDeletePropt();
 
   const hasLocation =
     sample.attrs.location?.latitude || sample.attrs.location?.longitude;
 
-  const startPointLocation: LatLngExpression = [
-    sample.attrs.location?.latitude,
-    sample.attrs.location?.longitude,
-  ];
   const endPointLocation = sample.attrs.location?.shape?.coordinates?.length
-    ? sample.attrs.location?.shape?.coordinates?.slice().reverse()
+    ? sample.attrs.location?.shape?.coordinates?.slice()
     : null;
 
   const finishPosition =
+    endPointLocation &&
+    endPointLocation[endPointLocation.length - 1].slice().reverse();
+
+  const startPosition =
     endPointLocation && endPointLocation[0].slice().reverse();
+
   const isFinished = sample.metadata.saved || sample.isTimerFinished();
   const isShapePolyline = sample.attrs.location?.shape?.type === 'LineString';
   const showStartMarker = isShapePolyline && hasLocation;
   const showLastMarker = isShapePolyline && endPointLocation;
-  const lastIcon = isFinished ? finishFlag : lastFlag;
 
-  const attachOnMoveWrap = () => {
-    const attachOnMove = () => {
-      if (!map) return;
-      const onMoveEnd = (): void => {
-        const { lat, lng } = map.getCenter();
+  const centerToLastTrackPoint = () => {
+    if (isFinished) return;
 
-        setMapCurrentCenter([lat, lng]);
-      };
+    const DEFAULT_LOCATED_ZOOM = 12;
 
-      map.on('moveend', onMoveEnd);
-    };
-    attachOnMove();
+    if (!Number.isFinite(location?.longitude)) return;
 
-    return function cleanup() {
-      if (!map) return;
+    const currentZoom = mapRef.current?.getZoom();
 
-      map.off('moveend');
-    };
+    const zoom =
+      currentZoom! > DEFAULT_LOCATED_ZOOM ? currentZoom : DEFAULT_LOCATED_ZOOM;
+
+    mapRef.current?.flyTo({
+      center: [location!.longitude, location!.latitude],
+      zoom,
+      speed: 2,
+    });
+  };
+  useEffect(centerToLastTrackPoint, [
+    isFinished,
+    location?.latitude,
+    location?.longitude,
+  ]);
+
+  const startingPointMarker = showStartMarker && (
+    <CircleMarker
+      id="start"
+      latitude={startPosition[0]}
+      longitude={startPosition[1]}
+      paint={{
+        'circle-color': 'white',
+        'circle-opacity': 1,
+        'circle-stroke-color': 'white',
+      }}
+    />
+  );
+
+  const finishPointMarker = showLastMarker && isFinished && (
+    <CircleMarker
+      id="finish"
+      latitude={finishPosition[0]}
+      longitude={finishPosition[1]}
+      paint={{
+        'circle-color': 'black',
+        'circle-stroke-color': 'white',
+      }}
+    />
+  );
+
+  const currentPointMarker = showLastMarker && !isFinished && (
+    <CircleMarker
+      id="finish"
+      latitude={finishPosition[0]}
+      longitude={finishPosition[1]}
+      paint={{
+        'circle-color': '#df9100',
+        'circle-opacity': 1,
+        'circle-stroke-color': '#df9100',
+      }}
+    />
+  );
+
+  const onShapeChange = async (shape: any) => {
+    if (!shape) {
+      const shouldDelete = await shouldDeleteShape();
+      if (!shouldDelete) return false;
+    }
+
+    setLocation(shape);
+    return true;
   };
 
-  useEffect(attachOnMoveWrap, [map]);
+  const toggleFavourites = () => setShowPastLocations(!showPastLocations);
+
+  const { isLocating, centerMapToCurrentLocation } =
+    useCenterMapToCurrentLocation(GPS);
 
   return (
     <Main className={`${isGPSTracking ? 'GPStracking' : ''}`}>
-      <IonToolbar id="area-edit-toolbar">
-        <IonTitle slot="start">{areaPretty}</IonTitle>
-      </IonToolbar>
-
       <MapContainer
-        tap={disableTapForIOS}
-        zoom={DEFAULT_ZOOM}
-        whenCreated={setTileLayer}
+        ref={mapRef}
+        accessToken={config.map.mapboxApiKey}
+        mapStyle="mapbox://styles/mapbox/satellite-streets-v10"
+        maxPitch={0}
+        initialViewState={location}
+        onMoveEnd={updateMapCentre}
       >
-        {showStartMarker && (
-          <Marker position={startPointLocation} icon={startFlag} />
-        )}
+        {!isDisabled && <Favourites.Control onClick={toggleFavourites} />}
 
-        {showLastMarker && <Marker position={finishPosition} icon={lastIcon} />}
-
-        <MapInfo
-          sample={sample}
-          isDisabled={isDisabled}
-          setLocation={setLocation}
-          shape={shape}
-          match={match}
+        <MapContainer.Control.Geolocate
+          isLocating={isLocating}
+          onClick={centerMapToCurrentLocation}
         />
 
-        {!isGPSTracking && !isDisabled && (
-          <MapControl position="topleft" className="favorite-btn">
-            <button
-              ref={buttonRef}
-              className="geolocate-btn"
-              onClick={(e: any) => {
-                L.DomEvent.preventDefault(e);
-                L.DomEvent.stopPropagation(e);
-                e.preventDefault();
-                e.stopPropagation();
+        <MapDraw shape={location?.shape} onChange={onShapeChange}>
+          {!isDisabled && <MapDraw.Control line polygon />}
 
-                setShowPastLocations(true);
-              }}
-            >
-              <IonIcon icon={starOutline} mode="md" size="large" />
-            </button>
-          </MapControl>
-        )}
+          <MapDraw.Context.Consumer>
+            {({ isEditing }: any) => {
+              if (isEditing) return null;
+              return (
+                <>
+                  <MapContainer.Marker {...location} />
+                  {startingPointMarker}
+                  {finishPointMarker}
+                  {currentPointMarker}
+                </>
+              );
+            }}
+          </MapDraw.Context.Consumer>
+        </MapDraw>
+
+        <Records sample={sample} />
       </MapContainer>
 
-      <IonModal
-        id="bottom-sheet-tracks"
+      <Favourites
         isOpen={showPastLocations}
-        backdropDismiss={false}
-        backdropBreakpoint={0.5}
-        breakpoints={SNAP_POSITIONS}
-        initialBreakpoint={DEFAULT_SNAP_POSITION}
-        canDismiss
-        onIonModalWillDismiss={() => {
-          setShowPastLocations(false);
-        }}
-      >
-        <IonHeader class="ion-no-border">
-          <IonToolbar />
-        </IonHeader>
-        <IonContent>
-          <PastLocationsList
-            onSelect={onSelectPastLoaction}
-            position={currentMapCenter}
-          />
-        </IonContent>
-      </IonModal>
+        sample={sample}
+        currentLocation={mapCenter}
+        onClose={() => setShowPastLocations(false)}
+      />
     </Main>
   );
 };
