@@ -21,6 +21,7 @@ export {
   Options as OccurrenceOptions,
   Metadata as OccurrenceMetadata,
 } from '@flumens/ionic/dist/models/Indicia/Occurrence';
+export { validateRemoteModel } from '@flumens/ionic/dist/models/Indicia/helpers';
 export {
   default as DrupalUserModel,
   Attrs as DrupalUserModelAttrs,
@@ -34,6 +35,7 @@ export { default as RouteWithModels } from '@flumens/ionic/dist/components/Route
 export { default as LongPressButton } from '@flumens/ionic/dist/components/LongPressButton';
 export { default as InputWithValidation } from '@flumens/ionic/dist/components/InputWithValidation';
 export { default as CounterInput } from '@flumens/ionic/dist/components/CounterInput';
+export { default as RadioInput } from '@flumens/ionic/dist/components/RadioInput';
 export { default as Main } from '@flumens/ionic/dist/components/Main';
 export { default as Header } from '@flumens/ionic/dist/components/Header';
 export { default as InfoMessage } from '@flumens/ionic/dist/components/InfoMessage';
@@ -44,6 +46,7 @@ export {
   Props as PageProps,
   AttrPropsExtended,
 } from '@flumens/ionic/dist/components/AttrPage';
+export { default as ModelValidationMessage } from '@flumens/ionic/dist/components/ModelValidationMessage';
 export { default as Gallery } from '@flumens/ionic/dist/components/Gallery';
 export { default as MenuAttrToggle } from '@flumens/ionic/dist/components/MenuAttrToggle';
 export { default as ModalHeader } from '@flumens/ionic/dist/components/ModalHeader';
@@ -56,10 +59,17 @@ export {
   MenuProps as MenuAttrItemFromModelMenuProps,
 } from '@flumens/ionic/dist/components/MenuAttrItemFromModel';
 export { default as UserFeedbackRequest } from '@flumens/ionic/dist/components/UserFeedbackRequest';
-export { default as ModelLocationMap } from '@flumens/ionic/dist/components/ModelLocationMap';
+export {
+  default as MapContainer,
+  useMapStyles,
+} from '@flumens/ionic/dist/components/Map/Container';
+export { default as MapHeader } from '@flumens/ionic/dist/components/Map/Header';
+export { default as CircleMarker } from '@flumens/ionic/dist/components/Map/Container/LocationMarker/CircleMarker';
+export { default as MapSettingsPanel } from '@flumens/ionic/dist/components/Map/SettingsPanel';
+export { default as MapDraw } from '@flumens/ionic/dist/components/Map/Draw';
+export * from '@flumens/ionic/dist/components/Map/utils';
 export { default as LongPressFabButton } from '@flumens/ionic/dist/components/LongPressFabButton';
 export { useToast, useAlert, useLoader } from '@flumens/ionic/dist/hooks';
-export * from '@flumens/ionic/dist/utils/validation';
 export * from '@flumens/ionic/dist/utils/image';
 export * from '@flumens/ionic/dist/utils/errors';
 export { default as date } from '@flumens/ionic/dist/utils/date';
